+++ ts/packages/knowPro/src/index.ts
@@ -17,3 +17,4 @@ export * from "./dateTimeSchema.js";
 export * from "./searchSchema.js";
 export * from "./searchTranslator.js";
 export * from "./textLocationIndex.js";
+export * from "./messageIndex.js";
